Day 14 part 2
Cute puzzle. Needed some exploration and guessing at what the tree
might look like. First try would be assuming that it would have a
vertical trunk section, so I looked for 10 robots in a line. Turns
out the image has a border that matches that.

Cleaned up the code a little after finding the answer to exit on
the tick it finds the tree, rather than continuing and needing to
look through the console manually for a tree